# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1"
  title: Acuris Example API
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /v1
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /reviews/{reviewId}:
    # binds a127 app logic to a route
    x-swagger-router-controller: review
    get:
      description: Returns a specific review by ID
      # used as the method name of the controller
      operationId: getReviewById
      parameters:
        - name: reviewId
          in: path
          description: The ID of the review
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/IndividualReviewResponse"
        "404":
          description: Could not find review
          schema:
            $ref: "#/definitions/ErrorResponse"  
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /countries: 
    x-swagger-router-controller: country
    get: 
        description: Retuns a list of countries
        operationId: listCountries
        responses: 
          "200":
            description: Success
            schema:
              $ref: "#/definitions/EnumResponse"
  /reviews:
    x-swagger-router-controller: review
    get:
      description: Returns paginated list of reviews
      operationId: listReviews
      parameters:
        - name: after
          in: query
          required: false
          type: string
        - name: before
          in: query
          required: false
          type: string
        - name: size
          in: query
          required: false
          type: number
        - name: countries
          in: query
          required: false
          type: string
          collectionFormat: csv
        - name: q
          in: query
          required: false
          type: string
        - name: fields
          in: query
          required: false
          type: string
          collectionFormat: csv
        - name: sort
          in: query
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ListOfReviewsResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  ListOfReviewsResponse:
    required:
      - results
    properties:
      approximateTotalCount:
        type: number
      results:
        type: array
        items:
          $ref: "#/definitions/IndividualReviewResponse"
  IndividualReviewResponse:
    required:
      - id
    properties:
      id:
        type: string
      country:
        type: string
      province:
        type: string
      variety:
         type: string
      kind:
        type: string
      price:
        type: string
      geography:
        type: string
      description:
        type: string
      designation:
        type: string
      winery:
        type: string
      region:
        type: string
      points:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  EnumResponse: 
    required:
      - name
    properties:
      name:
        type: string